예매를 위한 컨트롤러 세팅
