+++ src/main/java/com/ezen/biz/service/ThboardService.java
@@ -1,0 +1,9 @@
+package com.ezen.biz.service;
+
+import com.ezen.biz.dto.ThboardVO;
+
+public interface ThboardService {
+
+	ThboardVO seatinfo(ThboardVO vo);
+
+}
