+++ src/main/resources/mappings/thboard-mapping.xml
@@ -2,15 +2,15 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
 <mapper namespace="ThMapper">
 
-	<!-- <select id="adminCheck" resultType="String">
-		SELECT pwd FROM admin
+	<select id="seatinfo" resultType="Thboard">
+		SELECT * FROM seat
 		WHERE
 		id=#{id}
 	</select>
 
-	<select id="getAdmin" resultType="admin">
-		SELECT * FROM admin
+	 <select id="seatList" resultType="Thboard">
+		SELECT * FROM seat
 		WHERE
-		id=#{id}
-	</select> -->
+		thnum= #{thnum}
+	</select> 
 </mapper>
